# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  version: 1.0.0
  description: API documentation for the E-commerce platform
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        type:
          type: string
          enum: [OK, ERROR]
        message:
          type: string
        data:
          type: object
        meta:
          type: object
...
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
          format: uri
        website:
          type: string
          format: uri
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /brands:
    get:
      tags:
        - Brands
      summary: Get all brands (optionally filtered by search term)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: searchTerm
          schema:
            type: string
          description: Optional search term to filter brands by name (case insensitive)
      responses:
        '200':
          description: Brands retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    type: object
                    properties:
                      brands:
                    type: array
                    items:
                      $ref: '#/components/schemas/Brand'
        '500':
          description: Internal server error (e.g. "Failed to retrieve brands")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags:
        - Brands
      summary: Create a new brand
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Brand name (required, unique, min 2, max 50 chars)
                description:
                  type: string
                  description: Brand description (optional, max 1000 chars)
                logo:
                  type: object
                  properties:
                    url:
                      type: string
                      description: Logo URL (optional, must be a valid http(s) URL)
                    alt:
                      type: string
                      description: Alt text (optional)
                website:
                  type: string
                  description: Website URL (optional, must be a valid http(s) URL)
                contact:
                  type: object
                  properties:
                    email:
                      type: string
                      description: Contact email (optional, must be a valid email)
                    phone:
                      type: string
                      description: Contact phone (optional)
                    address:
                      type: object
                      properties:
                        street:
                          type: string
                        city:
                          type: string
                        state:
                          type: string
                        country:
                          type: string
                        zipCode:
                          type: string
                socialMedia:
                  type: object
                  properties:
                    facebook:
                      type: string
                    twitter:
                      type: string
                    instagram:
                      type: string
                    linkedin:
                      type: string
                orderPlaced:
                  type: array
                  items:
                    type: object
                    properties:
                      orderId:
                        type: string
                        required: true
                      status:
                        type: string
                        enum: [pending, processing, completed, cancelled]
                        default: pending
                      deliveryDate:
                        type: string
                        format: date-time
                      notes:
                        type: string
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            productId:
                              type: string
                              required: true
                            quantity:
                              type: number
                              required: true
                              minimum: 1
                            price:
                              type: number
                              required: true
                              minimum: 0
                      totalAmount:
                        type: number
                        required: true
                        minimum: 0
                outOfStockOrders:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                        required: true
                      quantity:
                        type: number
                        required: true
                        minimum: 1
                      status:
                        type: string
                        enum: [pending, submitted, received, cancelled]
                        default: pending
                      submittedAt:
                        type: string
                        format: date-time
                      expectedDeliveryDate:
                        type: string
                        format: date-time
                      notes:
                        type: string
                isActive:
                  type: boolean
                  default: true
                  description: Whether the brand is active (optional)
                displayOrder:
                  type: number
                  default: 0
                  description: Display order (optional)
                seo:
                  type: object
                  properties:
                    title:
                      type: string
                    description:
                      type: string
                    keywords:
                      type: array
                      items:
                        type: string
      responses:
        '201':
          description: Brand created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
        '500':
          description: Internal server error (e.g. "Failed to create brand")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags:
        - Brands
      summary: Update a brand (by ID)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the brand to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Brand name (optional, unique, min 2, max 50 chars)
                description:
                  type: string
                  description: Brand description (optional, max 1000 chars)
                logo:
                  type: object
                  properties:
                    url:
                      type: string
                      description: Logo URL (optional, must be a valid http(s) URL)
                    alt:
                      type: string
                      description: Alt text (optional)
                website:
                  type: string
                  description: Website URL (optional, must be a valid http(s) URL)
                contact:
                  type: object
                  properties:
                    email:
                      type: string
                      description: Contact email (optional, must be a valid email)
                    phone:
                      type: string
                      description: Contact phone (optional)
                    address:
                      type: object
                      properties:
                        street:
                          type: string
                        city:
                          type: string
                        state:
                          type: string
                        country:
                          type: string
                        zipCode:
                          type: string
                socialMedia:
                  type: object
                  properties:
                    facebook:
                      type: string
                    twitter:
                      type: string
                    instagram:
                      type: string
                    linkedin:
                      type: string
                orderPlaced:
                  type: array
                  items:
                    type: object
                    properties:
                      orderId:
                        type: string
                        required: true
                      status:
                        type: string
                        enum: [pending, processing, completed, cancelled]
                        default: pending
                      deliveryDate:
                        type: string
                        format: date-time
                      notes:
                        type: string
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            productId:
                              type: string
                              required: true
                            quantity:
                              type: number
                              required: true
                              minimum: 1
                            price:
                              type: number
                              required: true
                              minimum: 0
                      totalAmount:
                        type: number
                        required: true
                        minimum: 0
                outOfStockOrders:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                        required: true
                      quantity:
                        type: number
                        required: true
                        minimum: 1
                      status:
                        type: string
                        enum: [pending, submitted, received, cancelled]
                        default: pending
                      submittedAt:
                        type: string
                        format: date-time
                      expectedDeliveryDate:
                        type: string
                        format: date-time
                      notes:
                        type: string
                isActive:
                  type: boolean
                  default: true
                  description: Whether the brand is active (optional)
                displayOrder:
                  type: number
                  default: 0
                  description: Display order (optional)
                seo:
                  type: object
                  properties:
                    title:
                      type: string
                    description:
                      type: string
                    keywords:
                      type: array
                      items:
                        type: string
      responses:
        '200':
          description: Brand updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
        '404':
          description: Brand not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error (e.g. "Failed to update brand")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
        - Brands
      summary: Delete a brand (by ID)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the brand to delete
      responses:
        '200':
          description: Brand deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Brand not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error (e.g. "Failed to delete brand")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /brands/{id}:
    get:
      tags:
        - Brands
      summary: Get a single brand (by ID)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the brand to fetch
      responses:
        '200':
          description: Brand retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
        '404':
          description: Brand not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error (e.g. "Failed to retrieve brand")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...